Fix myprop, formatting in edo & issue in proba/prostoch
Fix the multiplicity of myprop in eplcommon; fixed some more formatting
in edo (didn't take the class though so can't say that much about the
math); found an issue in probastat and prostoch with the \N command.
Since I redefined it earlier to denote the set of natural numbers, it
became a doubly-defined command which LaTeX can't deal with. The issue
might reappear in other documents, will have to scourge the repo to find
out...