[LINMA1702] Notes: fix errors in first 4 cm's
The first four cm's had small errors related to punctuation,
that are fixed in this commit.
Namely, sentence-like punctuation in some of the programs
is now better spaced
(according to the style used in Bertsimas & Tsitsiklis).

A couple of mistakes were also corrected,
such as the meaning of 's.t.'
on the second line of a linear program.

Finally, some small typsetting errors in formulae
were also corrected.