[LINGI1101] Summary: add CM1 (chap. 1&2 in notes)
Create and add the contents of the first lecture to the summary.
This corresponds to the first and second chapters in the lecture notes,
even though the sections on truth tables and logical equivalence
still need to be written.
Will add those as soon as the lecture covers them
(presumably CM2).

Signed-off-by: Gilles Peiffer <gilles.peiffer@outlook.com>