Remove useless folder
[LINGI1341] Summary: start datalink layer

Add the subsection on framing to the datalink layer.
Also fixed some mistakes in the previous text.

[LINGI1341] Summary: add to datalink layer

Added a couple subsubsections
to the subsection about the datalink layer.
Namely, the subsections on
reliable transfer on top of an imperfect link,
and ABP.
Made all capital letter acronyms with \textsc
because I think it looks better.

Signed-off-by: Gilles Peiffer <gilles.peiffer@outlook.com>

[LINGI1341] Summary: finish S1

[LINGI1341] Summary: make window size note clearer

[LINGI1341] summary: add intro building a network

[LINGI1341] summary2: add cm2

[LINGI1341] summary2: add cm3

[LINGI1341] summary2: add cm4+5 and cm6 up to tcp

[LINGI1341] summary2: finish cm6, add cm 7 and cm8

[LINGI1341] summary2: add cm9

[LINGI1341] summary2: finish cm10